MarketPlace/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
pytest-django=4.7.0
